Migrate to String Web

+++ src/main/java/com/zavialov/repository/PostRepositoryImpl.java
@@ -1,10 +1,14 @@
 package com.zavialov.repository;
 
+import com.zavialov.exception.NotFoundException;
 import com.zavialov.model.Post;
+import org.springframework.stereotype.Repository;
 
 import java.util.List;
 import java.util.Optional;
 import java.util.concurrent.CopyOnWriteArrayList;
+
+@Repository
 
 public class PostRepositoryImpl implements PostRepository {
     List<Post> repository = new CopyOnWriteArrayList<>();
@@ -28,8 +32,11 @@ public class PostRepositoryImpl implements PostRepository {
         Post result = null;
         if (post.getId() == 0) {
             postCount++;
-            repository.add(new Post(postCount, post.getContent()));
+            post.setId(postCount);
+            repository.add(post);
             result = post;
+        } else if ((getById(post.getId()).isEmpty())) {
+            throw new NotFoundException();
         } else {
             for (Post value : repository) {
                 if (value.getId() == post.getId()) {
@@ -41,7 +48,12 @@ public class PostRepositoryImpl implements PostRepository {
         return result;
     }
 
-    public void removeById(long id) {
-        repository.removeIf(value -> value.getId() == id);
+    public String removeById(long id) {
+        if (getById(id).isEmpty()) {
+            throw new NotFoundException();
+        } else {
+            repository.removeIf(value -> value.getId() == id);
+            return "Post #" + id + " was successfully deleted.";
+        }
     }
 }
